refactor: move call_api to services.api
