+++ app/services/__init__.py
@@ -1,3 +1,3 @@
-from .api import api_client
+from .api import api_client, call_api
 
-__all__ = ["api_client"]
+__all__ = ["api_client", "call_api"]
